Remove unnecessary Sapling key types.
ZeWIF should not parse or otherwise decompose types for which there is a
protocol-defined encoding. Doing so would require ZeWIF to be updated
any time a protocol-defined encoding changes, and while in the case of
protocol additions this may sometimes be necessary anyway, it is
desirable to limit the complexity of those changes. Any wallet
interacting with the ZeWIF format will internally have the ability to
parse the standard encodings, and so reproducing the parsing logic in
ZeWIF is both duplicative and makes the ZeWIF encoding more brittle.