Remove component parts from unified address representation.
Clients of this library will necessarily have the ability to decompose
unified addresses. In addition to receiver parts, unified addresses may
include additional metadata items, and unified address encodings may
vary between Revision 0 and Revision 1 unified addresses. As a
consequence, working with the individual components of unified addresses
should be left as the responsibility of these downstream crates; this
also insulates the ZeWIF format from future changes to UAs, such as
those that may be associated with the introduction of ZSAs.