Add Sapling extended full viewing keys.
These are required in order to be able to properly represent imported
extended full viewing keys. However, in general attaching keys to to the
`sapling::Address` type is not ideal; this could result in the ZeWIF
storing a lot of redundant data in the case that the wallet holds many
diversified addresses. These keys will will likely be moved elsewhere in
a future commit.