Move transparent-only types to a `transparent` module & delete `SpendingKey`.
The `SpendingKey` type was not a good abstraction, in that it attempted
to cover spending key material for distinct protocols; its use inside of
the `TransparentSpendAuthority` type meant that it was possible to
construct a `TransparentSpendAuthority` using a Sapling spending key.
This change separates out transparent spending keys as a separate type.