Add `Cargo.lock` to ensure consistency of CI runs.
It is not uncommon for Rust projects to increase their MSRV without
updating to a new semver version; unfortunately, that then breaks CI
builds. To avert that, we include the `Cargo.lock` file in the
repository. It is not used in library builds.

In addition, having the Cargo.lock file present makes it easier to see
when dependencies are updated when reviewing changes.