Move binary parsing framework to `zewif-zcashd`.
The parsing framework that was initially introduced in this crate was
specialized to the `zcashd`-specific encodings of types. It is more
appropriate for this parsing code to live in the `zewif-zcashd` crate,
so as to avoid more general use of these formats in the Zcash ecosystem.

Along with the removal of the parsing framework, this completes the
removal of the `u160, u252, and u256` types, which are also
entirely zcashd-specific.