Remove incorrect components of `SaplingSentOutput` and `OrchardSentOutput`
There is no reason to store the individual components of a sent note
note; these components can either be recovered using the account's OVK,
or, if a separate OVK was used, then that should be stored
independently. If the output was created with an OVK policy that forbids
recovery of the note plaintext, then the output should not be retained
by the wallet in any case.