Remove the `TransactionStatus` type.
The transaction status as of the time of wallet export is generally not
reliable from the perspective of a wallet importing the exported
transaction; as such, it is up to the importing wallet to make its own
determinations about transaction status in the same way that it would
have intrinsically determined transaction status.

Generally, it's the responsibility of wallet export tools to construct
the wallet export format, and not to attempt to infer semantic
information about the exported data; the wallet export tool doesn't have
enough context to make these determinations.